Move Tags component in another file
